# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Provision Civils - Construction Job Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    put:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /backups/{id}/restore:
    post:
      operationId: restoreBackup
      tags: [backups]
      summary: Restore from a backup (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Restore initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
...
  /crash-reports:
    post:
      operationId: createCrashReport
      tags: [crash]
      summary: Submit a crash report from the mobile app
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrashReportInput"
      responses:
        "200":
          description: Report received
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
    get:
      operationId: listCrashReports
      tags: [crash]
      summary: List crash reports (admin only)
      parameters:
        - in: query
          name: resolved
          schema: { type: boolean }
      responses:
        "200":
          description: Crash report list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CrashReport"
  /crash-reports/{id}/resolve:
    post:
      operationId: resolveCrashReport
      tags: [crash]
      summary: Mark crash report as resolved (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Resolved
...
components:
  schemas:
...
    Client:
      type: object
      required: [id, companyName, createdAt]
      properties:
        id:
          type: integer
        companyName:
          type: string
        contactPerson:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        vatNumber:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ClientInput:
      type: object
      required: [companyName]
      properties:
        companyName:
          type: string
        contactPerson:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        vatNumber:
          type: string
        notes:
          type: string
...
    CrashReport:
      type: object
      required: [id, errorMessage, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        appVersion:
          type: ["string", "null"]
        platform:
          type: ["string", "null"]
        deviceInfo:
          nullable: true
        errorMessage:
          type: string
        stackTrace:
          type: ["string", "null"]
        extraContext:
          nullable: true
        resolved:
          type: boolean
        createdAt:
          type: string
    CrashReportInput:
      type: object
      required: [errorMessage]
      properties:
        appVersion:
          type: string
        platform:
          type: string
        deviceInfo:
          type: object
          additionalProperties: true
        errorMessage:
          type: string
        stackTrace:
          type: string
        extraContext:
          type: object
          additionalProperties: true
...